refactor(hooks): split wallet context into domain data hooks
Replace broad `useWalletData` usage with feature-specific hooks across UI
components (e.g., debt/credit, portfolio, notifications), including
separating portfolio state from NEPSE market data access.

This improves separation of concerns, reduces unnecessary coupling to the
global wallet context, and makes data dependencies clearer and easier to
maintain.refactor(hooks): split wallet context into domain data hooks

Replace broad `useWalletData` usage with feature-specific hooks across UI
components (e.g., debt/credit, portfolio, notifications), including
separating portfolio state from NEPSE market data access.

This improves separation of concerns, reduces unnecessary coupling to the
global wallet context, and makes data dependencies clearer and easier to
maintain.
